requirements.txt
flask
python-dotenv
cx_Oracle
